Fix so that a team cant be deleted if it has matches, and minor fixes

+++ src/main/java/se/torgammelgard/service/impl/TeamServiceImpl.java
@@ -1,13 +1,13 @@
 package se.torgammelgard.service.impl;
 
 import java.security.Principal;
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import se.torgammelgard.exception.TeamOwnsMatchesException;
 import se.torgammelgard.exception.UserNotFoundException;
 import se.torgammelgard.persistence.entities.Team;
 import se.torgammelgard.persistence.entities.User;
@@ -26,39 +26,47 @@ public class TeamServiceImpl implements TeamService {
     private UserRepository userRepository;
 
     @Override
-    public List<Team> findAllFor(Principal principal) throws UserNotFoundException {
-    	List<Team> teams = new ArrayList<>();
-    	User owner = userRepository.findByUsername(principal.getName());
-    	if (owner == null) {
-    		throw new UserNotFoundException();
-    	}
-    	for (Team team : teamRepository.findAll()) {
-    		if (team.getOwner() == owner) {
-    			teams.add(team);
-    		}
-    	}
-    	return teams;
+    public List<Team> findAllBelongingTo(Principal principal) throws UserNotFoundException {
+    	User user = findUser(principal);
+    	return teamRepository.findAllTeamsBelongingTo(user);
     }
     
     @Override
-    public Team save(Team team, Principal principal) throws UserNotFoundException {
+    public List<Team> findAllTeamsWithMatches(Principal principal) throws UserNotFoundException {
+    	User user = findUser(principal);
+    	return teamRepository.findAllTeamsWithMatches(user);
+    }
+    
+    @Override
+    public Team saveAndFlush(Team team, Principal principal) throws UserNotFoundException {
         if (team == null)
             return null;
-        User user = userRepository.findByUsername(principal.getName());
-        if (user == null) {
-        	throw new UserNotFoundException();
-        }
+        User user = findUser(principal);
         team.setOwner(user);
-        return teamRepository.save(team);
+        return teamRepository.saveAndFlush(team);
     }
 
     @Override
-    public void delete(Long id) {
-    	teamRepository.delete(id);
+    public void delete(Long id, Principal principal) throws UserNotFoundException, TeamOwnsMatchesException {
+    	User user = findUser(principal);
+    	List<Team> teams = teamRepository.findAllTeamsWithMatches(user);
+    	if (teams.isEmpty()) {
+    		teamRepository.delete(id);
+    	} else {
+    		throw new TeamOwnsMatchesException();
+    	}
     }
     
     @Override
     public Team find(Long id) {
         return teamRepository.findOne(id);
+    }
+    
+    private User findUser(Principal principal) throws UserNotFoundException {
+    	User user = userRepository.findByUsername(principal.getName());
+    	if (user == null) {
+    		throw new UserNotFoundException();
+    	}
+    	return user;
     }
 }
